backend/pyproject.toml
[project]
name = "turbo-potato-backend"
version = "0.1.0"
description = "Python FastAPI backend for Turbo Potato"
requires-python = ">=3.13"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "uvicorn>=0.30.0,<1.0.0",
  "pydantic>=2.9.0,<3.0.0",
  "pydantic-settings>=2.5.0,<3.0.0",
  "PyJWT[crypto]>=2.9.0,<3.0.0",
  "pyyaml>=6.0.2,<7.0.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.27.0,<1.0.0",
  "pytest>=8.3.0,<9.0.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
